Crate libpressio_sys

Structs

pressio

\file pressio.h \brief Pressio compressor loader

pressio_compressor

\file \brief Compress, decompress, and configure pressio and lossless compressors

pressio_data

\file \brief an abstraction for a contagious memory region of a specified type

pressio_io

\file \brief a generic io interface for libpressio

pressio_metrics

\file

pressio_option

pressio_options

\file \brief An iterator for a set of options for a compressor

pressio_options_iter

Constants

pressio_conversion_safety_pressio_conversion_explicit

all of the above, and conversions that are explicitly convertible with an explicit cast in C++, see also std::is_convertable

pressio_conversion_safety_pressio_conversion_implicit

conversions that are implicitly convertible without a narrowing conversion see also [dcl.init.list] in the C++ standard

pressio_conversion_safety_pressio_conversion_special

all of the above, and conversions that require a special function call (i.e. atoi) if this function fails, NULL will be returned

pressio_dtype_pressio_bool_dtype

8 bit data type

pressio_dtype_pressio_byte_dtype

8 bit data type

pressio_dtype_pressio_double_dtype

64 bit double precision floating point

pressio_dtype_pressio_float_dtype

32 bit double precision floating point

pressio_dtype_pressio_int8_dtype

8 bit signed integer

pressio_dtype_pressio_int16_dtype

16 bit signed integer

pressio_dtype_pressio_int32_dtype

32 bit signed integer

pressio_dtype_pressio_int64_dtype

64 bit signed integer

pressio_dtype_pressio_uint8_dtype

8 bit unsigned integer

pressio_dtype_pressio_uint16_dtype

16 bit unsigned integer

pressio_dtype_pressio_uint32_dtype

32 bit unsigned integer

pressio_dtype_pressio_uint64_dtype

64 bit unsigned integer

pressio_option_type_pressio_option_bool_type

option is a boolean

pressio_option_type_pressio_option_charptr_array_type

option is an array of c-style strings

pressio_option_type_pressio_option_charptr_type

option is a non-owning pointer to a c-style string

pressio_option_type_pressio_option_data_type

option is a pressio_data structure

pressio_option_type_pressio_option_double_type

option is a 64 bit double precision floating point

pressio_option_type_pressio_option_dtype_type

option is a pressio_dtype

pressio_option_type_pressio_option_float_type

option is a 32 bit single precision floating point

pressio_option_type_pressio_option_int8_type

option is a 8 bit signed integer

pressio_option_type_pressio_option_int16_type

option is a 16 bit signed integer

pressio_option_type_pressio_option_int32_type

option is a 32 bit signed integer

pressio_option_type_pressio_option_int64_type

option is a 64 bit signed integer

pressio_option_type_pressio_option_threadsafety_type

option is a threadsafety

pressio_option_type_pressio_option_uint8_type

option is a 8 bit unsigned integer

pressio_option_type_pressio_option_uint16_type

option is a 16 bit unsigned integer

pressio_option_type_pressio_option_uint32_type

option is a 32 bit unsigned integer

pressio_option_type_pressio_option_uint64_type

option is a 64 bit unsigned integer

pressio_option_type_pressio_option_unset_type

option is a non-owning pointer to a arbitrary data

pressio_option_type_pressio_option_userptr_type

option is a non-owning pointer to a arbitrary data

pressio_options_key_status_pressio_options_key_does_not_exist

the requested key does not exist

pressio_options_key_status_pressio_options_key_exists

the requested key exists but is not set, for pressio_option_set_* functions indicates a type mismatch

pressio_options_key_status_pressio_options_key_set

the requested key exists and is set, evaluates to false

pressio_thread_safety_pressio_thread_safety_multiple

calls are safe if and only if only one thread will execute the above sequences of calls to an instance of the compressor at a time

pressio_thread_safety_pressio_thread_safety_serialized

calls are safe if and only if only one thread will execute the above sequences of calls to any instance of the compressor at a time

pressio_thread_safety_pressio_thread_safety_single

use of this compressor in a multi-threaded environment is unsafe.

Functions

pressio_compressor_check_options⚠

Validates that only defined options have been set. This can be useful for programmer errors. This function should NOT be used with any option structure which contains options for multiple compressors. Other checks MAY be preformed implementing compressors.

pressio_compressor_clone⚠

Clones a compressor and its configuration including metrics information

pressio_compressor_compress⚠

compresses the data in data using the specified compressor \param[in] compressor compressor to be used \param[in] input data to be compressed and associated metadata \param[in,out] output when passed in, \c output MAY contain metadata (type, dimentions) and additionally MAY contain a buffer. pressio_data_free will be called on the pointer passed into this function if a new owning pressio_data structure is returned. when passed out, \c output will contain either: 1. The same pointer passed in to \c output if the compressor supports using a provided buffer for the results of the compression and contains has a buffer. It is recommended the user provide an owning pressio_data structure if passing a pressio_data structure with a buffer. 2. A new owning pressio_data structure with the results of the compression \returns 0 if successful, positive values on errors, negative values on warnings

pressio_compressor_compress_many⚠

compress multiple data buffers in one api call. Underlying implementations may do this in serial or parallel

pressio_compressor_decompress⚠

decompresses the compressed data using the specified compressor calling this without calling libpressio_compressor_set_options() has undefined behavior decompressing with a compressor with different settings than used for compression has undefined behavior

pressio_compressor_decompress_many⚠

decompress multiple data buffers in one api call. Underlying implementations may do this in serial or parallel

pressio_compressor_error_code⚠

\param[in] compressor the compressor to query \returns last error code for the compressor

pressio_compressor_error_msg⚠

\param[in] compressor the compressor to query \returns last error message for the compressor

pressio_compressor_get_configuration⚠

\returns a pressio options struct that represents the compile time configuration of the compressor \param[in] compressor which compressor to get compile-time configuration for

pressio_compressor_get_documentation⚠

\returns a pressio options struct that represents the documentation for the compressor \param[in] compressor which compressor to get documentation for

pressio_compressor_get_metrics⚠

\param[in] compressor the compressor to get the metrics plugin for \returns the current pressio_metrics* structure

pressio_compressor_get_metrics_results⚠

\param[in] compressor the compressor to get results from \returns a pressio_options structure containing the metrics returned by the provided metrics plugin \see libpressio_metricsplugin for how to compute results

pressio_compressor_get_name⚠

Get the name of a compressor \param[in] compressor the compressor to get the name of \returns a string with the compressor name. The string becomes invalid if the name is set_name is called.

pressio_compressor_get_options⚠

\returns a pressio options struct that represents the current options of the compressor \param[in] compressor which compressor to get options for

pressio_compressor_get_prefix⚠

Returns the name this compressor uses its keys

pressio_compressor_major_version⚠

\param[in] compressor the compressor to query \returns the major version number or 0 if there is none

pressio_compressor_metrics_get_options⚠

Gets the options for a metrics structure \param[in] compressor the metrics structure to get options for \returns a new pressio_options structure with the options for the metrics

pressio_compressor_metrics_set_options⚠

Gets the options for a metrics structure \param[in] compressor the compressor structure to get metrics options for \param[in] options the options to set \returns 0 if successful, positive values on errors, negative values on warnings

pressio_compressor_minor_version⚠

\param[in] compressor the compressor to query \returns the major version number or 0 if there is none

pressio_compressor_patch_version⚠

\param[in] compressor the compressor to query \returns the major version number or 0 if there is none

pressio_compressor_release⚠

\param[in] compressor deallocates a reference to a compressor.

pressio_compressor_set_metrics⚠

\param[in] compressor the compressor to set metrics plugin for \param[in] plugin the configured libpressio_metricsplugin plugin to use

pressio_compressor_set_name⚠

Assign a new name to a compressor. Names are used to prefix options in meta-compressors.

pressio_compressor_set_options⚠

sets the options for the pressio_compressor. Compressors MAY choose to ignore some subset of options passed in if there valid but conflicting settings (i.e. two settings that adjust the same underlying configuration). Additionally, if one of the two settings is a generic one (i.e. pressio:abs) the compressor specific version should prevail. Compressors SHOULD return an error value if configuration failed due to a missing required setting or an invalid one. Users MAY call pressio_compressor_error_msg() to get more information about the warnings or errors Compressors MUST ignore any and all options set that they do not support.

pressio_compressor_version⚠

Get the version and feature information. The version string may include more information than major/minor/patch provide. \param[in] compressor the compressor to query \returns a implementation specific version string

pressio_data_cast⚠

Creates a data buffer with the specified type from an existing buffer

pressio_data_copy⚠

allocates a buffer and then copies the underlying memory to the new buffer \param[in] data the pressio data to copy from \param[out] out_bytes the number of bytes that were copied

pressio_data_domain_id⚠

returns the domain_id of the data

pressio_data_dtype⚠

\param[in] data the pressio data to query \returns an integer code corresponding to the data-type

pressio_data_free⚠

frees a pressio_data structure, but not the underlying data \param[in] data data structure to free

pressio_data_get_bytes⚠

returns the number of bytes to represent the data \param[in] data the pressio data to query \returns the number of bytes to represent the data

pressio_data_get_capacity_in_bytes⚠

returns the number of bytes of capacity \param[in] data the pressio data to query \returns the number of bytes to represent the data

pressio_data_get_dimension⚠

returns the value of a given dimension. \param[in] data the pressio data to query \param[in] dimension zero indexed dimension \returns the dimension or 0 If the dimension requested exceeds num_dimensions

pressio_data_has_data⚠

\param[in] data the pressio data to query \returns an integer code corresponding to the data-type

pressio_data_libc_free_fn⚠

a custom deleter that uses libc’s free and ignores the metadata \param[in] data to be deallocated with free() \param[in] metadata ignored

pressio_data_new_clone⚠

allocates a new pressio_data structure and corresponding data and copies data from provided data structure \param[in] src the pressio_data structure to be cloned \returns the newly allocated copy

pressio_data_new_copy⚠

allocates a new pressio_data structure and corresponding data and copies data in from the specified source use this function when the underlying data pointer may be deleted

pressio_data_new_empty⚠

allocates a new pressio_data without data.

pressio_data_new_move⚠

allocates a new pressio_data structure using data that was already allocated. data referenced by this structure will be deallocated using the function deleter when the pressio_data structure is freed.

pressio_data_new_nonowning⚠

allocates a new pressio_data structure, it does NOT take ownership of data.

pressio_data_new_nonowning_domain⚠

allocates a new pressio_data structure, it does NOT take ownership of data.

pressio_data_new_nonowning_from_data⚠

allocates a new pressio_data structure, it does NOT take ownership of data.

pressio_data_new_owning⚠

allocates a new pressio_data structure and corresponding data. The corresponding data is uninitialized

pressio_data_new_owning_from_data⚠

allocates a new pressio_data structure, it does NOT take ownership of data.

pressio_data_num_dimensions⚠

\param[in] data the pressio data to query \returns the number of dimensions contained in the object

pressio_data_num_elements⚠

returns the total number of elements to represent the data \param[in] data the pressio data to query \returns the total number of elements to represent the data

pressio_data_ptr⚠

non-owning access to the first element of the raw data \param[in] data the pressio data to query \param[out] out_bytes the number of bytes that follow this pointer (ignored if NULL is passed) \returns a non-owning type-pruned pointer to the first element of the raw data \see pressio_data_dtype to get the data-type

pressio_data_reshape⚠

modifies the dimensions of a pressio_data structure inplace.

pressio_data_select⚠

Copies a possibly strided subset of from the data pointer

pressio_data_transpose⚠

Transposes a data-buffer

pressio_dtype_is_floating⚠

\returns non-zero if the type is a floating point value

pressio_dtype_is_numeric⚠

\returns non-zero if the type is a numeric value

pressio_dtype_is_signed⚠

\returns non-zero if the type is signed

pressio_dtype_size⚠

\returns the size in bytes of a libpressio recognized type

pressio_error_code⚠

\param[in] library the pointer to the library \returns a machine-readable error code for the last error on the library object

pressio_error_msg⚠

\param[in] library the pointer to the library \returns a human-readable error message for the last error on the library object

pressio_features⚠

\returns a string containing all the features supported by this version separated by a space. Some features are compressors, but not all are. \see pressio_get_compressor the compressor_ids may be passed to pressio_get_compressor

pressio_get_compressor⚠

\param[in] library the pointer to the library \param[in] compressor_id the compressor to use \returns non-owning pointer to the requested instantiated pressio compressor; it may return the same pointer on multiple calls \see pressio_features for a list of available compressors

pressio_get_io⚠

\param[in] library the name of the io module \param[in] io_module the name of the io module \returns a new handle to an io module

pressio_instance⚠

gets a reference to a new instance of libpressio; initializes the library if necessary \returns a pointer to a library instance

pressio_io_check_options⚠

Validates that only defined options have been set. This can be useful for programmer errors. This function should NOT be used with any option structure which contains options for multiple io modules. Other checks MAY be preformed implementing io modules.

pressio_io_clone⚠

Creates a new reference to an io module with the same configuration

pressio_io_error_code⚠

\param[in] io the io module to query \returns last error code for the io

pressio_io_error_msg⚠

\param[in] io the io module to query \returns last error message for the io

pressio_io_free⚠

\param[in] io deallocates a reference to a io module.

pressio_io_get_configuration⚠

\returns a pressio options struct that represents the compile time configuration of the io module \param[in] io which io to get compile-time configuration for

pressio_io_get_documentation⚠

\returns a pressio options struct that represents the documentation of the io module \param[in] io which io to get documentation for

pressio_io_get_name⚠

Get the name of a io \param[in] io the io to get the name of \returns a string with the io name. The string becomes invalid if the name is set_name is called.

pressio_io_get_options⚠

\returns a pressio options struct that represents the current options of the io module \param[in] io which io to get options for

pressio_io_major_version⚠

\param[in] io the io module to query \returns the major version number or 0 if there is none

pressio_io_minor_version⚠

\param[in] io the io module to query \returns the major version number or 0 if there is none

pressio_io_patch_version⚠

\param[in] io the io module to query \returns the major version number or 0 if there is none

pressio_io_read⚠

reads a pressio_data buffer from some persistent storage \param[in] io the object to preform the read \param[in] data data object to populate, or nullptr to allocate it from the file if supported by the plugin. Data passed to this call should be considered “moved” in a c++11 sense. \returns nullptr on failures, non-null on success

pressio_io_read_many⚠

read multiple buffers

pressio_io_set_name⚠

Assign a new name to a io. Names are used to prefix options in meta-ios.

pressio_io_set_options⚠

sets the options for the pressio_io. io modules MAY choose to ignore some subset of options passed in if there valid but conflicting settings (i.e. two settings that adjust the same underlying configuration). io modules SHOULD return an error value if configuration failed due to a missing required setting or an invalid one. Users MAY call pressio_io_error_msg() to get more information about the warnings or errors io modules MUST ignore any and all options set that they do not support.

pressio_io_version⚠

Get the version and feature information. The version string may include more information than major/minor/patch provide. \param[in] io the io to query \returns a implementation specific version string

pressio_io_write⚠

write a pressio_data buffer from some persistent storage \param[in] io the object to preform the write \param[in] data data object to write \returns 0 if successful, positive values on errors, negative values on warnings

pressio_io_write_many⚠

write multiple buffers

pressio_major_version⚠

\returns the major version of the library

pressio_metrics_clone⚠

Clones a pressio_metrics object and its configuration \param[in] metrics the metrics object to clone \returns a new reference to a pressio metrics object

pressio_metrics_error_code⚠

\param[in] metrics the metrics to query \returns last error code for the metrics

pressio_metrics_error_msg⚠

\param[in] metrics the metrics to query \returns last error message for the metrics

pressio_metrics_evaluate⚠

Evaluate a metric that operates on a data buffer and return a corresponding options structure

pressio_metrics_free⚠

frees a metrics structure \param[in] metrics the metrics structure to free

pressio_metrics_get_configuration⚠

Gets the configuration for a metrics structure \param[in] metrics the metrics structure to get configuration for \returns a new pressio_options structure with the configuration for the metrics

pressio_metrics_get_documentation⚠

Gets the documentation for a metrics structure \param[in] metrics the metrics structure to get documentation for \returns a new pressio_options structure with the documentation for the metrics

pressio_metrics_get_name⚠

Get the name of a metrics \param[in] metrics the metrics to get the name of \returns a string with the metrics name. The string becomes invalid if the name is set_name is called.

pressio_metrics_get_options⚠

Gets the options for a metrics structure \param[in] metrics the metrics structure to get options for \returns a new pressio_options structure with the options for the metrics

pressio_metrics_get_results⚠

Gets the results from a metrics structure \param[in] metrics the metrics structure to get results from \returns a new pressio_options structure the metrics structure to get results from

pressio_metrics_set_name⚠

Assign a new name to a metrics. Names are used to prefix options in meta-metrics.

pressio_metrics_set_options⚠

Gets the options for a metrics structure \param[in] metrics the metrics structure to get options for \param[in] options the to set \returns 0 if successful, positive values on errors, negative values on warnings

pressio_minor_version⚠

\returns the minor version of the library

pressio_new_metric⚠

creates a new metrics structure

pressio_new_metrics⚠

creates a composite metrics structure

pressio_option_as_set⚠

Sets pressio_option for the specified key with a generic pressio_options preserving the type of the key in the options structure using an implicit cast if necessary \param[in] lhs the option to get an pressio_option for \param[in] rhs value to assign in the pressio_options structure \returns pressio_options_key_set if the lhs is modified, pressio_options_key_exists otherwise

pressio_option_cast_set⚠

Sets pressio_option for the specified key with a generic pressio_options preserving the type of the key in the options structure using the specified cast if necessary \param[in] lhs the option to get an pressio_option for \param[in] rhs value to assign in the pressio_options structure \param[in] safety what kind of conversions to allow \returns pressio_options_key_set if the lhs is modified, pressio_options_key_exists otherwise

pressio_option_convert⚠

converts between one type and another \param[in] option the option to convert \param[in] type the type to convert to \param[in] safety how safe to make perform a conversion \returns a new option value of the type specified if possible, otherwise returns NULL.

pressio_option_convert_implicit⚠

\param[in] option the option to convert \param[in] type the type to convert to \returns a new option value of the type specified if possible, otherwise returns NULL. \see pressio_option_convert behaves as if this function was called if with safety=pressio_conversion_implicit

pressio_option_free⚠

frees the memory associated with a returned option \param[in] options the option to free

pressio_option_get_bool⚠

pressio_option_get_data⚠

Gets the value stored in the pressio_option. Calling this with the improper dtype has undefined behavior \param[in] option the option to retrieve a value from \returns a new owning copy of the value contained in the option

pressio_option_get_double⚠

pressio_option_get_dtype⚠

pressio_option_get_float⚠

pressio_option_get_integer⚠

pressio_option_get_integer8⚠

pressio_option_get_integer16⚠

pressio_option_get_integer64⚠

pressio_option_get_string⚠

pressio_option_get_strings⚠

Gets the value stored in the pressio_option. Calling this with the improper dtype has undefined behavior \param[in] option the option to retrieve a value from \param[out] size the size of the array returned \returns the value contained in the option, the returned value must be freed

pressio_option_get_threadsafety⚠

pressio_option_get_type⚠

gets the type of the returned pressio_option \param[in] option the option to get the dtype for \returns the type the option contains

pressio_option_get_uinteger⚠

pressio_option_get_uinteger8⚠

pressio_option_get_uinteger16⚠

pressio_option_get_uinteger64⚠

pressio_option_get_userptr⚠

pressio_option_has_value⚠

returns true if the option contains a value \param[in] option the option to check for a value \return true if the option contains a value

pressio_option_new⚠

creates an empty pressio_option \returns a new option structure with dtype pressio_option_unset

pressio_option_new_bool⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_data⚠

Creates a new pressio_option containing the specified value \param[in] data the value to use to create the object \returns a pointer to a new pressio option set to value passed in

pressio_option_new_double⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_dtype⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_float⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_integer⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_integer8⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_integer16⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_integer64⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_string⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_strings⚠

Creates a new pressio_option containing the specified value \param[in] values the value to use to create the object \param[in] size the length of the array \returns a pointer to a new pressio option set to value passed in

pressio_option_new_threadsafety⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_uinteger⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_uinteger8⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_uinteger16⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_uinteger64⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_userptr⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object
\returns a pointer to a new pressio option set to value passed in\

pressio_option_new_userptr_managed⚠

Creates a new pressio_option containing the specified value \param[in] value the value to use to create the object \param[in] metadata to use to manage the allocation of value \param[in] deleter deletes the value \param[in] copy copies the value \returns a pointer to a new pressio option set to value passed in

pressio_option_set_bool⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_data⚠

Sets the option to an integer value \param[in] option the option to set \param[in] value the value to set

pressio_option_set_double⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_dtype⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_float⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_integer⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_integer8⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_integer16⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_integer64⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_string⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_strings⚠

Sets the option to an integer value \param[in] option the option to set \param[in] values the value to set \param[in] size the size of the array of values

pressio_option_set_threadsafety⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_type⚠

set the type of the passed pressio_option \param[in] option the option to get the dtype for \param[in] type the option to get the dtype for

pressio_option_set_uinteger⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_uinteger8⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_uinteger16⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_uinteger64⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_userptr⚠

Sets the option to an integer value
\param[in] option the option to set
\param[in] value the value to set \

pressio_option_set_userptr_managed⚠

Set a pressio_option containing the specified value \param[in] option the option structure to set \param[in] value the value to use to create the object \param[in] metadata to use to manage the allocation of value \param[in] deleter deletes the value \param[in] copy copies the value

pressio_option_to_string⚠

Create a human readable string for the option passed.

pressio_options_as_bool⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_data⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_double⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_dtype⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_float⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_integer⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_integer8⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_integer16⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_integer64⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_set⚠

Sets pressio_option for the specified key with a generic pressio_options preserving the type of the key in the options structure using an implicit cast if necessary \param[in] options the option to get an pressio_option for \param[in] key the key to get from the options structure \param[in] option value to assign in the pressio_options structure \returns pressio_options_key_set if the lhs is modified, pressio_options_key_exists if the key exists but can’t convert, and pressio_options_key_does_not_exist if the key does not exist;

pressio_options_as_string⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_strings⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] size the number of strings returned. 0 on error \param[out] values the value retrieved, only if it is convertible. both the values and the pointer must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_threadsafety⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_uinteger⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_uinteger8⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_uinteger16⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_uinteger64⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_as_userptr⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_bool⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_data⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_double⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_dtype⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_float⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_integer⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_integer8⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_integer16⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_integer64⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_set⚠

Sets pressio_option for the specified key with a generic pressio_options preserving the type of the key in the options structure using the specified cast if necessary \param[in] options the option to get an pressio_option for \param[in] key the key to get from the options structure \param[in] option value to assign in the pressio_options structure \param[in] safety what kind of conversions to allow \returns pressio_options_key_set if the lhs is modified, pressio_options_key_exists if the key exists but can’t convert, and pressio_options_key_does_not_exist if the key does not exist;

pressio_options_cast_string⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_strings⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] size the number of strings returned. 0 on error \param[out] values the value retrieved, only if it is convertible. both the values and the pointer must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_threadsafety⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_uinteger⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_uinteger8⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_uinteger16⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_uinteger64⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_cast_userptr⚠

Gets an particular key in an options structure, casting it if necessary \param[in] options the options structure to modify \param[in] key the key to change \param[in] safety what kind of conversions to allow \param[out] value the value retrieved, only if it is convertible. If returning a char*, the memory must be freed with free() \returns a status code \see pressio_options_key_status status codes returned

pressio_options_clear⚠

Clear the value associated with a key, but retains the entry in the options.

pressio_options_copy⚠

Copies the memory associated with this pressio option structure

pressio_options_exists⚠

\param[in] options the option to get an pressio_options_key_status for \param[in] key the key to get from the options structure \returns a key status for the requested key

pressio_options_free⚠

Frees the memory associated with a pressio option structure

pressio_options_get⚠

Gets a generic pressio_option for the specified key. Calling this with an nonexistent key has undefined behavior \param[in] options the option to get an pressio_option for \param[in] key the key to get from the options structure \returns a new copy of the corresponding pressio_option

pressio_options_get_bool⚠

Gets a particular value in a map if it exists

pressio_options_get_data⚠

Gets a particular value in a map if it exists

pressio_options_get_double⚠

Gets a particular value in a map if it exists

pressio_options_get_dtype⚠

Gets a particular value in a map if it exists

pressio_options_get_float⚠

Gets a particular value in a map if it exists

pressio_options_get_integer⚠

Gets a particular value in a map if it exists

pressio_options_get_integer8⚠

Gets a particular value in a map if it exists

pressio_options_get_integer16⚠

Gets a particular value in a map if it exists

pressio_options_get_integer64⚠

Gets a particular value in a map if it exists

pressio_options_get_iter⚠

Get an iterator over all keys the use of any of the pressio_options_set* functions invalidates the iterator

pressio_options_get_string⚠

Gets a particular value in a map if it exists

pressio_options_get_strings⚠

Gets a particular value in a map if it exists

pressio_options_get_threadsafety⚠

Gets a particular value in a map if it exists

pressio_options_get_uinteger⚠

Gets a particular value in a map if it exists

pressio_options_get_uinteger8⚠

Gets a particular value in a map if it exists

pressio_options_get_uinteger16⚠

Gets a particular value in a map if it exists

pressio_options_get_uinteger64⚠

Gets a particular value in a map if it exists

pressio_options_get_userptr⚠

Gets a particular value in a map if it exists

pressio_options_iter_free⚠

\param[in] iter the iterator to free frees memory associated with the iterator

pressio_options_iter_get_key⚠

get the key for the current position of the iterator \param[in] iter the iterator to get the current value for \returns a non-owning pointer to the key name associated with the current position of the iterator

pressio_options_iter_get_value⚠

get the value for the current position of the iterator \param[in] iter the iterator to get the current value for \returns a owning pointer to the value associated with the current position of the iterator

pressio_options_iter_has_value⚠

Returns true if the current position has a value \param[in] iter to check if it has a next value \returns true if the current iterator position has a value

pressio_options_iter_next⚠

advances the iterator to the next entry \param[in] iter the iterator to advance

pressio_options_merge⚠

Merges two pressio options together into one. Copies all keys and corresponding values from rhs not in lhs into a new structure.

pressio_options_new⚠

Creates an empty pressio_options structure returns NULL if the allocation fails \returns a pointer to the new options structure \see pressio_options_free

pressio_options_num_set⚠

\param[in] options the options structure to get the size of \returns the number of keys with the status pressio_options_key_set

pressio_options_set⚠

Sets pressio_option for the specified key with a generic pressio_options \param[in] options the option to get an pressio_option for \param[in] key the key to get from the options structure \param[in] option value to set in the pressio_options structure

pressio_options_set_bool⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_data⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_double⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_dtype⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_float⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_integer⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_integer8⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_integer16⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_integer64⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_string⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_strings⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] size the number of strings passed \param[in] values the value to change to

pressio_options_set_threadsafety⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_type⚠

Sets pressio_option to the specified type \param[in] options the option to set type of a pressio_option for \param[in] key the key to get from the options structure \param[in] type value to set in the pressio_options structure

pressio_options_set_uinteger⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_uinteger8⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_uinteger16⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_uinteger64⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_userptr⚠

Sets an particular key in an options structure with the given key to a value \param[in] options the options structure to modify \param[in] key the key to change \param[in] value the value to change to

pressio_options_set_userptr_managed⚠

Creates a new pressio_option containing the specified value \param[in] options the options structure to set \param[in] key the value to use to create the object \param[in] value the value to use to create the object \param[in] metadata to use to manage the allocation of value \param[in] deleter deletes the value \param[in] copy copies the value

pressio_options_size⚠

\param[in] options the options structure to get the size of \returns the number of keys with either the status pressio_options_key_set or pressio_options_key_exists

pressio_options_to_string⚠

Create a human readable string for the options passed.

pressio_patch_version⚠

\returns the patch version of the library

pressio_register_all⚠

ensure that all libpressio components are initialized, this is called by pressio_instance

pressio_release⚠

\param[in] library the pointer to the library informs the library that this instance is no longer required; the pointer passed becomes invalid

pressio_supported_compressors⚠

\returns a string containing all the compressors supported by this version separated by a space \see pressio_get_supported_compressors the compressor_ids may be passed to pressio_get_compressor

pressio_supported_io_modules⚠

\returns the list supported io_modules seperated by whitespace

pressio_supported_metrics⚠

\returns a string containing all the metrics supported by this version separated by a space

pressio_version⚠

it will not return more information than the tailored functions below \returns a string with version and feature information

Type Aliases

pressio_conversion_safety

level of safety to require for conversions

pressio_data_delete_fn

signature for a custom deleter for pressio_data

pressio_dtype

data types recognized by libpressio for compression and decompression

pressio_option_type

possible types contained in a pressio_option, more types may be added in the future

pressio_options_key_status

possible status of a particular key in the option structure

pressio_thread_safety

reports the level of thread safety supported by the compressor.