libpressio_sys

Function pressio_compressor_decompress

pub unsafe extern "C" fn pressio_compressor_decompress(
    compressor: *mut pressio_compressor,
    input: *const pressio_data,
    output: *mut pressio_data,
) -> c_int

decompresses the compressed data using the specified compressor calling this without calling libpressio_compressor_set_options() has undefined behavior decompressing with a compressor with different settings than used for compression has undefined behavior

\param[in] compressor compressor to be used \param[in] input data to be decompressed and associated metadata \param[in,out] output when passed in, \c output SHOULD contain the metadata (type, dimentions) for the output of the compression if available. pressio_data_free will be called the pointer passed in during this function. when passed out, it will contain either 1. The same pointer passed in to \c output if the compressor supports using a provided buffer for the results of the compression and contains has a buffer. It is recommended the user provide an owning pressio_data structure if passing a pressio_data structure with a buffer. 2. A new owning pressio_data structure with the results of the compression \returns 0 if successful, positive values on errors, negative values on warnings \see pressio_data_new_empty often used as the pointer passed into \c output \see pressio_data_new_move often used as the pointer passed out of \c output