libpressio_sys

Function pressio_data_transpose

pub unsafe extern "C" fn pressio_data_transpose(
    data: *const pressio_data,
    axis: *const usize,
) -> *mut pressio_data

Transposes a data-buffer

\param[in] data the data to copy from \param[in] axis the order of the dimensions if null is chosen the dimensions are reversed

\returns a new pressio data structure containing a transposed copy if an error occurs, an new empty structure is returned instead.