libpressio_sys

Function pressio_metrics_evaluate

pub unsafe extern "C" fn pressio_metrics_evaluate(
    metrics: *mut pressio_metrics,
    uncompressed: *const pressio_data,
    compressed: *const pressio_data,
    decompressed: *const pressio_data,
) -> *mut pressio_options

Evaluate a metric that operates on a data buffer and return a corresponding options structure

If the metric does not support being called on data buffers, the returned object is undefined.

\param[in] metrics the metrics object to invoke \param[in] uncompressed the data before compression, if nullptr, begin/end_compress_impl will not be called \param[in] compressed the data after compression, if nullptr, the metrics object MAY choose not to compute some or all metrics, but SHOULD compute as many as possible \param[in] decompressed the data after decompression, if nullptr, begin/end_decompress_impl will not be called

\returns a new pressio_options as if pressio_metrics_get_results was called.